Separate hex util from other utils
