+++ yggdrash-node/src/main/java/io/yggdrash/node/api/AdminApiImpl.java
@@ -5,6 +5,7 @@ import com.google.gson.JsonObject;
 import com.googlecode.jsonrpc4j.spring.AutoJsonRpcServiceImpl;
 import io.yggdrash.common.config.DefaultConfig;
 import io.yggdrash.common.crypto.HashUtil;
+import io.yggdrash.common.crypto.HexUtil;
 import io.yggdrash.common.util.ByteUtil;
 import io.yggdrash.common.util.FileUtil;
 import io.yggdrash.common.util.JsonUtil;
@@ -259,8 +260,7 @@ public class AdminApiImpl implements AdminApi {
         this.body = JsonUtil.parseJsonArray(command.getBody());
 
         // body length check
-        long bodyLength = ByteUtil.byteArrayToLong(
-                Hex.decode(header.get("bodyLength").getAsString()));
+        long bodyLength = HexUtil.hexStringToLong(header.get("bodyLength").getAsString());
 
         if (body.toString().length() != bodyLength) {
             errorMsg.append(" BodyLength is not valid.");
@@ -268,8 +268,7 @@ public class AdminApiImpl implements AdminApi {
         }
 
         // timestamp check (3 min)
-        long timestamp = ByteUtil.byteArrayToLong(
-                Hex.decode(header.get("timestamp").getAsString()));
+        long timestamp = HexUtil.hexStringToLong(header.get("timestamp").getAsString());
         if (timestamp < System.currentTimeMillis() - (COMMAND_ACTIVE_TIME)) {
             log.error("Timestamp is not valid.");
             errorMsg.append(" Timestamp is not valid.");
